[DEV-5301] Requested changes and additional comments
