+++ usaspending_api/database_scripts/matview_generator/mv_grant_award_search.json
@@ -189,6 +189,9 @@
       "name": "period_of_performance_current_end_date",
       "columns": [{"name": "period_of_performance_current_end_date", "order": "DESC NULLS LAST"}]
     }, {
+      "name": "recipient_hash",
+      "columns": [{"name": "recipient_hash"}]
+    }, {
       "name": "recipient_name",
       "where": "recipient_name IS NOT NULL",
       "columns": [{"name": "recipient_name"}]
