+++ usaspending_api/database_scripts/matview_generator/mv_idv_award_search.json
@@ -191,6 +191,9 @@
       "name": "ordered_ordering_period_end_date",
       "columns": [{"name": "ordering_period_end_date", "order": "DESC NULLS LAST"}]
     }, {
+      "name": "recipient_hash",
+      "columns": [{"name": "recipient_hash"}]
+    }, {
       "name": "recipient_name",
       "where": "recipient_name IS NOT NULL",
       "columns": [{"name": "recipient_name"}]
