+++ usaspending_api/database_scripts/matview_generator/mv_pre2008_award_search.json
@@ -4,14 +4,15 @@
   "matview_sql": [
     "SELECT",
     "  to_tsvector(CONCAT_WS(' ',",
-    "    COALESCE(recipient_lookup.recipient_name, transaction_fpds.awardee_or_recipient_legal),",
+    "    COALESCE(recipient_lookup.recipient_name, transaction_fpds.awardee_or_recipient_legal, transaction_fabs.awardee_or_recipient_legal),",
     "    transaction_fpds.naics,",
     "    transaction_fpds.naics_description,",
     "    psc.description,",
     "    (SELECT string_agg(tn.description, ' ') FROM transaction_normalized AS tn WHERE tn.award_id = awards.id GROUP BY tn.award_id)",
     "  )) AS keyword_ts_vector,",
     "  to_tsvector(awards.piid) AS award_ts_vector,",
-    "  to_tsvector(COALESCE(recipient_lookup.recipient_name, transaction_fpds.awardee_or_recipient_legal)) AS recipient_name_ts_vector,",
+    "  to_tsvector(COALESCE(recipient_lookup.recipient_name, transaction_fpds.awardee_or_recipient_legal, transaction_fabs.awardee_or_recipient_legal)",
+    "  ) AS recipient_name_ts_vector,",
     "",
     "  tas.treasury_account_identifiers,",
     "",
@@ -37,12 +38,12 @@
     "",
     "  COALESCE(recipient_lookup.recipient_hash, MD5(UPPER(",
     "    CASE",
-    "      WHEN transaction_fpds.awardee_or_recipient_uniqu IS NOT NULL THEN CONCAT('duns-', transaction_fpds.awardee_or_recipient_uniqu)",
-    "      ELSE CONCAT('name-', transaction_fpds.awardee_or_recipient_legal) END",
+    "      WHEN COALESCE(transaction_fpds.awardee_or_recipient_uniqu, transaction_fabs.awardee_or_recipient_uniqu) IS NOT NULL THEN CONCAT('duns-', COALESCE(transaction_fpds.awardee_or_recipient_uniqu, transaction_fabs.awardee_or_recipient_uniqu))",
+    "      ELSE CONCAT('name-', COALESCE(transaction_fpds.awardee_or_recipient_legal, transaction_fabs.awardee_or_recipient_legal)) END",
     "  ))::uuid) AS recipient_hash,",
-    "  UPPER(COALESCE(recipient_lookup.recipient_name, transaction_fpds.awardee_or_recipient_legal)) AS recipient_name,",
-    "  transaction_fpds.awardee_or_recipient_uniqu AS recipient_unique_id,",
-    "  transaction_fpds.ultimate_parent_unique_ide AS parent_recipient_unique_id,",
+    "  UPPER(COALESCE(recipient_lookup.recipient_name, transaction_fpds.awardee_or_recipient_legal, transaction_fabs.awardee_or_recipient_legal)) AS recipient_name,",
+    "  COALESCE(transaction_fpds.awardee_or_recipient_uniqu, transaction_fabs.awardee_or_recipient_uniqu) AS recipient_unique_id,",
+    "  COALESCE(transaction_fpds.ultimate_parent_unique_ide, transaction_fabs.ultimate_parent_unique_ide) AS parent_recipient_unique_id,",
     "  latest_transaction.business_categories,",
     "",
     "  latest_transaction.action_date,",
@@ -113,7 +114,7 @@
     "    duns,",
     "    recipient_hash",
     "  FROM recipient_lookup AS rlv",
-    "  ) recipient_lookup ON recipient_lookup.duns = transaction_fpds.awardee_or_recipient_uniqu AND transaction_fpds.awardee_or_recipient_uniqu IS NOT NULL",
+    "  ) recipient_lookup ON recipient_lookup.duns = COALESCE(transaction_fpds.awardee_or_recipient_uniqu, transaction_fabs.awardee_or_recipient_uniqu) AND COALESCE(transaction_fpds.awardee_or_recipient_uniqu, transaction_fabs.awardee_or_recipient_uniqu) IS NOT NULL",
     "LEFT OUTER JOIN",
     "  psc ON (transaction_fpds.product_or_service_code = psc.code)",
     "LEFT OUTER JOIN",
@@ -207,6 +208,9 @@
     }, {
       "name": "ordered_ordering_period_end_date",
       "columns": [{"name": "ordering_period_end_date", "order": "DESC NULLS LAST"}]
+    }, {
+      "name": "recipient_hash",
+      "columns": [{"name": "recipient_hash"}]
     }, {
       "name": "recipient_name",
       "where": "recipient_name IS NOT NULL",
