+++ usaspending_api/disaster/v2/views/disaster_base.py
@@ -215,33 +215,3 @@ class LoansPaginationMixin(_BasePaginationMixin):
     def pagination(self):
         sortable_columns = ["id", "code", "description", "count", "obligation", "outlay", "face_value_of_loan"]
         return self.run_models(sortable_columns)
-
-
-class ElasticsearchSpendingPaginationMixin(_BasePaginationMixin):
-    sum_column_mapping = {"obligation": "total_covid_obligation", "outlay": "total_covid_outlay"}
-    sort_column_mapping = {
-        "description": "_key",
-        "count": "_count",
-        **sum_column_mapping,
-    }
-
-    @cached_property
-    def pagination(self):
-        return self.run_models(list(self.sort_column_mapping), default_sort_column="description")
-
-
-class ElasticsearchLoansPaginationMixin(_BasePaginationMixin):
-    sum_column_mapping = {
-        "obligation": "total_covid_obligation",
-        "outlay": "total_covid_outlay",
-        "face_value_of_loan": "total_loan_value",
-    }
-    sort_column_mapping = {
-        "description": "_key",
-        "count": "_count",
-        **sum_column_mapping,
-    }
-
-    @cached_property
-    def pagination(self):
-        return self.run_models(list(self.sort_column_mapping), default_sort_column="description")
