+++ usaspending_api/disaster/v2/views/elasticsearch_base.py
@@ -2,6 +2,7 @@ from abc import abstractmethod
 from typing import List, Optional, Dict
 
 from django.conf import settings
+from django.utils.functional import cached_property
 from elasticsearch_dsl import Q as ES_Q, A
 from rest_framework.request import Request
 from rest_framework.response import Response
@@ -10,14 +11,44 @@ from usaspending_api.awards.v2.lookups.lookups import loan_type_mapping
 from usaspending_api.common.cache_decorator import cache_response
 from usaspending_api.common.data_classes import Pagination
 from usaspending_api.common.elasticsearch.search_wrappers import AwardSearch
-from usaspending_api.common.exceptions import ElasticsearchConnectionException
+from usaspending_api.common.exceptions import ForbiddenException
 from usaspending_api.common.helpers.generic_helper import get_pagination_metadata
 from usaspending_api.common.query_with_filters import QueryWithFilters
-from usaspending_api.disaster.v2.views.disaster_base import DisasterBase
+from usaspending_api.disaster.v2.views.disaster_base import DisasterBase, _BasePaginationMixin
 from usaspending_api.search.v2.elasticsearch_helper import (
     get_scaled_sum_aggregations,
     get_number_of_unique_terms_for_awards,
 )
+
+
+class ElasticsearchSpendingPaginationMixin(_BasePaginationMixin):
+    sum_column_mapping = {"obligation": "total_covid_obligation", "outlay": "total_covid_outlay"}
+    sort_column_mapping = {
+        "description": "_key",
+        "count": "_count",
+        **sum_column_mapping,
+    }
+
+    @cached_property
+    def pagination(self):
+        return self.run_models(list(self.sort_column_mapping), default_sort_column="description")
+
+
+class ElasticsearchLoansPaginationMixin(_BasePaginationMixin):
+    sum_column_mapping = {
+        "obligation": "total_covid_obligation",
+        "outlay": "total_covid_outlay",
+        "face_value_of_loan": "total_loan_value",
+    }
+    sort_column_mapping = {
+        "description": "_key",
+        "count": "_count",
+        **sum_column_mapping,
+    }
+
+    @cached_property
+    def pagination(self):
+        return self.run_models(list(self.sort_column_mapping), default_sort_column="description")
 
 
 class ElasticsearchDisasterBase(DisasterBase):
@@ -86,12 +117,14 @@ class ElasticsearchDisasterBase(DisasterBase):
         }
         sort_values = {"order": {self.sort_column_mapping[self.pagination.sort_key]: self.pagination.sort_order}}
 
+        # As of writing this the value of settings.ES_ROUTING_FIELD is the only high cardinality aggregation that
+        # we support. Since the Elasticsearch clusters are routed by this field we don't care to get a count of
+        # unique buckets, but instead we use the upper_limit and don't allow an upper_limit > 10k.
         if self.agg_key == settings.ES_ROUTING_FIELD:
             size = self.pagination.upper_limit
             shard_size = size
             group_by_agg_key_values = {**sort_values}
             bucket_sort_values = {**pagination_values}
-
         else:
             if self.bucket_count == 0:
                 return None
@@ -102,8 +135,8 @@ class ElasticsearchDisasterBase(DisasterBase):
                 bucket_sort_values = {**pagination_values, **sort_values}
 
         if shard_size > 10000:
-            raise ElasticsearchConnectionException(
-                "Current filters return too many unique items. Narrow filters to return results."
+            raise ForbiddenException(
+                "Current filters return too many unique items. Narrow filters to return results or use downloads."
             )
 
         # Define all aggregations needed to build the response
